Add LoRA pass
