+++ examples/open_llama/README.md
@@ -80,7 +80,17 @@ LoRA adapters that were fine-tuned on the training dataset.
 
 The relevant config file is [llama_qlora.json](llama_qlora.json). It corresponds to the [guqnaco 7b example in the original qlora implementation](https://github.com/artidoro/qlora/blob/main/scripts/finetune_guanaco_7b.sh).
 
-Requirements file: [requirements-qlora.txt](requirements-qlora.txt)
+Requirements file: [requirements-lora.txt](requirements-lora.txt)
+
+### Fine-tune Open Llama Model on a code generation dataset using LoRA
+This workflow fine-tunes Open LLaMA model using [LoRA] to generate code given a prompt.
+
+The relevant config file is [open_llama_lora_tinycodes.json](open_llama_lora_tinycodes.json). The code language is set to `Python` but can be changed to other languages by changing the `language` field in the config file.
+Supported languages are Python, TypeScript, JavaScript, Ruby, Julia, Rust, C++, Bash, Java, C#, and Go. Refer to the [dataset card](https://huggingface.co/datasets/nampdn-ai/tiny-codes) for more details on the dataset.
+
+Note: You must be logged in to HuggingFace using `huggingface-cli login` to download the dataset or update `token` field in the config file with your HuggingFace token.
+
+Requirements file: [requirements-lora.txt](requirements-lora.txt)
 
 ### Fine-tune Open Llama Model on a code generation dataset using QLoRA
 This workflow fine-tunes Open LLaMA model using [QLoRA] to generate code given a prompt.
@@ -90,7 +100,7 @@ Supported languages are Python, TypeScript, JavaScript, Ruby, Julia, Rust, C++, 
 
 Note: You must be logged in to HuggingFace using `huggingface-cli login` to download the dataset or update `token` field in the config file with your HuggingFace token.
 
-Requirements file: [requirements-qlora.txt](requirements-qlora.txt)
+Requirements file: [requirements-lora.txt](requirements-lora.txt)
 
 ### Optimizing Open Llama Model with Azure Arc
 This workflow optimizes Open Llama model on Azure ML compute, and evaluate output models on your device. Please connect your device to Azure Arc by following instruction: [Self-hosted Kubernetes cluster](https://microsoft.github.io/Olive/tutorials/azure_arc.html)
