+++ examples/open_llama/open_llama_lora_tinycodes.json
@@ -37,12 +37,9 @@
         }
     },
     "passes": {
-        "qlora": {
-            "type": "QLoRA",
+        "lora": {
+            "type": "LoRA",
             "config": {
-                "compute_dtype": "bfloat16",
-                "quant_type": "nf4",
-                "double_quant": true,
                 "lora_r": 64,
                 "lora_alpha": 16,
                 "lora_dropout": 0.1,
@@ -51,12 +48,12 @@
                 "training_args": {
                     "seed": 0,
                     "data_seed": 42,
-                    "per_device_train_batch_size": 16,
-                    "per_device_eval_batch_size": 16,
+                    "per_device_train_batch_size": 8,
+                    "per_device_eval_batch_size": 8,
                     "gradient_accumulation_steps": 1,
                     "gradient_checkpointing": true,
                     "learning_rate": 0.0002,
-                    "max_steps": 1500,
+                    "max_steps": 500,
                     "logging_steps": 10,
                     "evaluation_strategy": "steps",
                     "eval_steps": 100,
@@ -79,8 +76,7 @@
                 "accelerators": ["gpu"]
             }
         },
-        "execution_providers": ["CPUExecutionProvider"],
         "cache_dir": "cache",
-        "output_dir" : "models/qlora-tiny-codes"
+        "output_dir" : "models/llama/lora"
     }
 }
