+++ olive/passes/pytorch/__init__.py
@@ -2,9 +2,9 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-from olive.passes.pytorch.qlora import QLoRA
+from olive.passes.pytorch.lora import LoRA, QLoRA
 from olive.passes.pytorch.quantization_aware_training import QuantizationAwareTraining
 from olive.passes.pytorch.sparsegpt import SparseGPT
 from olive.passes.pytorch.torch_trt_conversion import TorchTRTConversion
 
-__all__ = ["QLoRA", "QuantizationAwareTraining", "SparseGPT", "TorchTRTConversion"]
+__all__ = ["LoRA", "QLoRA", "QuantizationAwareTraining", "SparseGPT", "TorchTRTConversion"]
